% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Bayes Factors for Bayesact Class}
\usage{
bayes_factor(
  x1,
  x2,
  resp = NULL,
  newdata = NULL,
  sev_samples = NULL,
  sample_max = 1e+06,
  custom_pfun = NULL,
  ...
)
}
\description{
Bayes Factors for Bayesact Class
}
