% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_map.R
\docType{data}
\name{dfun_map}
\alias{dfun_map}
\title{Distribution Function Mapping}
\format{
An object of class \code{list} of length 15.
}
\usage{
dfun_map
}
\description{
Distribution Function Mapping
}
\keyword{datasets}
