% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{kfold}
\alias{kfold}
\title{K-Fold Cross Validation for Bayesact Class}
\usage{
kfold(
  x,
  resp = NULL,
  newdata = NULL,
  sev_samples = NULL,
  sample_max = 1e+06,
  custom_pfun = NULL,
  ...
)
}
\description{
K-Fold Cross Validation for Bayesact Class
}
