% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormpower.R
\name{lnormpower}
\alias{lnormpower}
\title{LogNormal-Power BRMS Custom Family}
\usage{
lnormpower(link = "identity", link_sigma = "log", link_alpha = "log1p")
}
\description{
LogNormal-Power BRMS Custom Family
}
