% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormpower.R
\docType{data}
\name{lnormpower_stan_funs}
\alias{lnormpower_stan_funs}
\title{LogNormal-Power BRMS Custom Family Stan Functions}
\format{
An object of class \code{character} of length 1.
}
\usage{
lnormpower_stan_funs
}
\description{
LogNormal-Power BRMS Custom Family Stan Functions
}
\keyword{datasets}
