% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormpower.R
\docType{data}
\name{lnormpower_stanvars}
\alias{lnormpower_stanvars}
\title{LogNormal-Power BRMS Custom Family Stan Variables}
\format{
An object of class \code{stanvars} of length 1.
}
\usage{
lnormpower_stanvars
}
\description{
LogNormal-Power BRMS Custom Family Stan Variables
}
\keyword{datasets}
