% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo}
\alias{loo}
\title{Leave One Out Criterion for Bayesact Class}
\usage{
loo(
  x,
  resp = NULL,
  newdata = NULL,
  sev_samples = NULL,
  sample_max = 1e+06,
  custom_pfun = NULL,
  ...
)
}
\description{
Leave One Out Criterion for Bayesact Class
}
